Refactored app environment
