+++ requirements.txt
@@ -1,11 +1,14 @@
 bcrypt==4.2.1
 blinker==1.9.0
+cachelib==0.13.0
 click==8.1.8
 dnspython==2.7.0
 Flask==3.1.0
 Flask-Login==0.6.3
+Flask-Session==0.8.0
 itsdangerous==2.2.0
 Jinja2==3.1.5
 MarkupSafe==3.0.2
+msgspec==0.18.6
 pymongo==4.10.1
 Werkzeug==3.1.3
